+++ practice/chapter01/statement.js
@@ -1,71 +1,7 @@
+import createStatementData from './createStatementData';
+
 export default function statement(invoice, plays) {
-  createStatementData(invoice, plays);
-}
-
-function createStatementData(invoice, plays) {
-  const statementData = {};
-  statementData.customer = invoice.customer;
-  statementData.performances = invoice.performances.map(enrichPerformance);
-  statementData.totalAmount = totalAmount(statementData);
-  statementData.totalVolumeCredits = totalVolumeCredits(statementData);
-  return renderPlainText(statementData, plays);
-
-  function enrichPerformance(aPerformance) {
-    const result = Object.assign({}, aPerformance);
-    result.play = playFor(result);
-    result.amount = amountFor(result);
-    result.volumeCredits = volumeCreditsFor(result);
-    return result;
-  }
-
-  function playFor(aPerformance) {
-    return plays[aPerformance.playId];
-  }
-
-  function amountFor(aPerformance) {
-    let result = 0;
-
-    switch (aPerformance.play.type) {
-      case 'tragedy':
-        result = 40_000;
-
-        if (aPerformance.audience > 30) {
-          result += 1_000 * (aPerformance.audience - 30);
-        }
-        break;
-      case 'comedy':
-        result = 30_000;
-
-        if (aPerformance.audience > 20) {
-          result += 10_000 + 500 * (aPerformance.audience - 20);
-        }
-        result += 300 * aPerformance.audience;
-        break;
-
-      default:
-        throw new Error(`알 수 없는 장르: ${aPerformance.play.type}`);
-    }
-
-    return result;
-  }
-
-  function volumeCreditsFor(perf) {
-    let result = 0;
-
-    result += Math.max(perf.audience - 30, 0);
-
-    if ('comedy' === perf.play.type) result += Math.floor(perf.audience / 5);
-
-    return result;
-  }
-
-  function totalAmount(data) {
-    return data.performances.reduce((total, p) => total + p.amount, 0);
-  }
-
-  function totalVolumeCredits(data) {
-    return data.performances.reduce((total, p) => total + p.volumeCredits, 0);
-  }
+  renderPlainText(createStatementData(invoice, plays));
 }
 
 function renderPlainText(data) {
